remove duplicate replace function
